refactor: drop Java-type leakage from AuthTokenManager public surface
The Java AuthTokenManager interface is impl-detail; client code should
target a duck-typed Ruby protocol (`get_token` + `handle_security_exception`)
that works identically under MRI and JRuby — same drop-in contract every
other public type follows.

- Move the Proc-based shortcut to shared
  (`Internal::AuthTokenManagers::Custom`); pure Ruby, no Java refs.
- Add an MRI surface (`Driver::AuthTokenManagers.custom`); `basic` /
  `bearer` raise NotImplementedError until managed-auth's retry
  semantics land on MRI.
- JRuby's `Ext::AuthTokenManagers.custom` returns the same shared
  Custom; basic/bearer keep delegating to the Java factory.
- New JRuby-only adapter (`Ext::Internal::AuthTokenManagerAdapter`)
  bridges any duck-typed manager to Java's AuthTokenManager interface
  inside `GraphDatabase.driver`. Dispatch switches on
  `respond_to?(:get_token)`, so a user-supplied class works without
  including the Java interface.
