+++ lib/jruby/neo4j/driver/ext/auth_token_managers.rb
@@ -5,11 +5,12 @@ module Neo4j
     module Ext
       # Mirrors Java's `AuthTokenManagers.basic(Supplier<AuthToken>)` /
       # `bearer(Supplier<AuthTokenAndExpiration>)` factories with Ruby
-      # keyword-arg surfaces, and adds a `custom` factory for the full
-      # `AuthTokenManager` interface (get_token + handle_security_exception)
-      # — the form testkit drives via `NewAuthTokenManager`. The Custom
-      # class implements the Java interface directly so JRuby can pass it
-      # to the driver wherever an `AuthTokenManager` is accepted.
+      # keyword-arg surfaces. The `custom` factory returns the shared
+      # duck-typed `Internal::AuthTokenManagers::Custom` (same class
+      # MRI uses); the boundary into Java's `AuthTokenManager`
+      # interface is crossed by `Internal::AuthTokenManagerAdapter`
+      # inside `GraphDatabase.driver`, so client code never names a
+      # Java type.
       module AuthTokenManagers
         def basic(supplier:)
           super(supplier)
@@ -20,26 +21,7 @@ module Neo4j
         end
 
         def custom(get_token:, handle_security_exception:)
-          Custom.new(get_token, handle_security_exception)
-        end
-
-        class Custom
-          include Java::OrgNeo4jDriver::AuthTokenManager
-
-          def initialize(get_token, handle_security_exception)
-            @get_token = get_token
-            @handle_security_exception = handle_security_exception
-          end
-
-          # CompletionStage<AuthToken> — our supplier is synchronous so
-          # we wrap the result in a pre-completed future.
-          def get_token
-            java.util.concurrent.CompletableFuture.completed_future(@get_token.call)
-          end
-
-          def handle_security_exception(token, exception)
-            @handle_security_exception.call(token, exception)
-          end
+          Neo4j::Driver::Internal::AuthTokenManagers::Custom.new(get_token, handle_security_exception)
         end
       end
     end
